Release 2.1 (Reformatted code)

+++ library/src/main/java/com/heinrichreimersoftware/materialdrawer/DrawerFrameLayout.java
@@ -56,8 +56,7 @@ public class DrawerFrameLayout extends DrawerLayout {
         int colorPrimaryDark = a.getColor(0, 0);
         if (colorPrimaryDark != 0) {
             setStatusBarBackgroundColor(colorPrimaryDark);
-        }
-        else{
+        } else {
             setStatusBarBackgroundColor(getResources().getColor(android.R.color.black));
         }
 
@@ -66,8 +65,7 @@ public class DrawerFrameLayout extends DrawerLayout {
         int drawerMaxWidth = a.getDimensionPixelSize(0, 0);
         if (drawerMaxWidth != 0) {
             setDrawerMaxWidth(drawerMaxWidth);
-        }
-        else{
+        } else {
             resetDrawerMaxWidth();
         }
         a.recycle();
@@ -84,7 +82,7 @@ public class DrawerFrameLayout extends DrawerLayout {
      *
      * @param theme Theme to set
      */
-    public DrawerFrameLayout setDrawerTheme(DrawerTheme theme){
+    public DrawerFrameLayout setDrawerTheme(DrawerTheme theme) {
         mDrawer.setDrawerTheme(theme);
         return this;
     }
@@ -92,7 +90,7 @@ public class DrawerFrameLayout extends DrawerLayout {
     /**
      * Resets the drawer theme
      */
-    public DrawerFrameLayout resetDrawerTheme(){
+    public DrawerFrameLayout resetDrawerTheme() {
         mDrawer.resetDrawerTheme();
         return this;
     }
@@ -100,7 +98,7 @@ public class DrawerFrameLayout extends DrawerLayout {
     /**
      * Gets the drawer theme
      */
-    public DrawerTheme getDrawerTheme(){
+    public DrawerTheme getDrawerTheme() {
         return mDrawer.getDrawerTheme();
     }
 
@@ -110,7 +108,7 @@ public class DrawerFrameLayout extends DrawerLayout {
      *
      * @param drawerMaxWidth Max drawer width to set
      */
-    public DrawerFrameLayout setDrawerMaxWidth(int drawerMaxWidth){
+    public DrawerFrameLayout setDrawerMaxWidth(int drawerMaxWidth) {
         mDrawer.setDrawerMaxWidth(drawerMaxWidth);
         return this;
     }
@@ -120,7 +118,7 @@ public class DrawerFrameLayout extends DrawerLayout {
      *
      * @param drawerMaxWidthResource Max drawer width resource to set
      */
-    public DrawerFrameLayout setDrawerMaxWidthResource(int drawerMaxWidthResource){
+    public DrawerFrameLayout setDrawerMaxWidthResource(int drawerMaxWidthResource) {
         mDrawer.setDrawerMaxWidthResource(drawerMaxWidthResource);
         return this;
     }
@@ -128,7 +126,7 @@ public class DrawerFrameLayout extends DrawerLayout {
     /**
      * Resets the max drawer width
      */
-    public DrawerFrameLayout resetDrawerMaxWidth(){
+    public DrawerFrameLayout resetDrawerMaxWidth() {
         mDrawer.resetDrawerMaxWidth();
         return this;
     }
@@ -136,7 +134,7 @@ public class DrawerFrameLayout extends DrawerLayout {
     /**
      * Gets the max drawer width
      */
-    public int getDrawerMaxWidth(){
+    public int getDrawerMaxWidth() {
         return mDrawer.getDrawerMaxWidth();
     }
 
@@ -356,7 +354,7 @@ public class DrawerFrameLayout extends DrawerLayout {
      * @param id The item ID
      * @return Item from the drawer
      */
-    public DrawerItem findItemById(int id) {
+    public DrawerItem findItemById(long id) {
         mDrawer.findItemById(id);
         return null;
     }
@@ -375,7 +373,7 @@ public class DrawerFrameLayout extends DrawerLayout {
      *
      * @return Position of the selected item
      */
-    public int getSelectedPosition(){
+    public int getSelectedPosition() {
         return mDrawer.getSelectedPosition();
     }
 
@@ -384,7 +382,7 @@ public class DrawerFrameLayout extends DrawerLayout {
      *
      * @param id The item ID
      */
-    public void selectItemById(int id) {
+    public void selectItemById(long id) {
         mDrawer.selectItemById(id);
     }
 
@@ -527,7 +525,7 @@ public class DrawerFrameLayout extends DrawerLayout {
      * @param id The item ID
      * @return Item from the drawer
      */
-    public DrawerItem findFixedItemById(int id) {
+    public DrawerItem findFixedItemById(long id) {
         mDrawer.findFixedItemById(id);
         return null;
     }
@@ -546,7 +544,7 @@ public class DrawerFrameLayout extends DrawerLayout {
      *
      * @return Position of the selected item
      */
-    public int getSelectedFixedPosition(){
+    public int getSelectedFixedPosition() {
         return mDrawer.getSelectedFixedPosition();
     }
 
@@ -555,7 +553,7 @@ public class DrawerFrameLayout extends DrawerLayout {
      *
      * @param id The item ID
      */
-    public void selectFixedItemById(int id) {
+    public void selectFixedItemById(long id) {
         mDrawer.selectFixedItemById(id);
     }
 
@@ -638,14 +636,14 @@ public class DrawerFrameLayout extends DrawerLayout {
     /**
      * Opens the drawer
      */
-    public void openDrawer(){
+    public void openDrawer() {
         openDrawer(mDrawer);
     }
 
     /**
      * Closes the drawer
      */
-    public void closeDrawer(){
+    public void closeDrawer() {
         closeDrawer(mDrawer);
     }
 }
